+++ sdk/include/opentelemetry/sdk/metrics/view/attributes_processor.h
@@ -11,6 +11,7 @@
 #include "opentelemetry/common/attribute_value.h"
 #include "opentelemetry/common/key_value_iterable.h"
 #include "opentelemetry/nostd/string_view.h"
+#include "opentelemetry/sdk/common/attribute_validity.h"
 #include "opentelemetry/sdk/common/custom_hash_equality.h"
 #include "opentelemetry/sdk/metrics/state/filtered_ordered_attribute_map.h"
 #include "opentelemetry/version.h"
@@ -59,7 +60,8 @@ public:
   MetricAttributes process(
       const opentelemetry::common::KeyValueIterable &attributes) const noexcept override
   {
-    MetricAttributes result(attributes);
+    MetricAttributes result(
+        opentelemetry::sdk::common::KeyValueFilterIterable(attributes, "[Metrics] "));
     return result;
   }
 
@@ -86,7 +88,9 @@ public:
       const opentelemetry::common::KeyValueIterable &attributes) const noexcept override
   {
     MetricAttributes result;
-    attributes.ForEachKeyValue(
+    opentelemetry::sdk::common::KeyValueFilterIterable validate_attributes{attributes,
+                                                                           "[Metrics] "};
+    validate_attributes.ForEachKeyValue(
         [&](nostd::string_view key, opentelemetry::common::AttributeValue value) noexcept {
           if (opentelemetry::sdk::common::find_heterogeneous(allowed_attribute_keys_, key) !=
               allowed_attribute_keys_.end())